add Footer & NavBar components and update index, docs, package.json
