+++ package.json
@@ -17,6 +17,7 @@
     "@coreui/react": "^4.4.0",
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
+    "@mui/icons-material": "^5.10.14",
     "@mui/material": "^5.10.6",
     "@mui/styled-engine-sc": "^5.10.6",
     "bootstrap": "^5.2.1",
